refactors seeds

+++ priv/repo/seeds.exs
@@ -10,21 +10,17 @@
 # We recommend using the bang functions (`insert!`, `update!`
 # and so on) as they will fail if something goes wrong.
 
-Reader.Repo.delete_all(Reader.Article)
+if Mix.env == :dev do
+  Reader.Repo.delete_all(Reader.Article)
 
-articles = [
-  %Reader.Article{url: "http://www.elixir.com", title: "A fantastic article about Elixir", category: "elixir"},
-  %Reader.Article{url: "http://www.elixir.org", title: "A fantastic article about Elixir", category: "elixir"},
-  %Reader.Article{url: "http://www.ruby.com", title: "A fantastic article about Ruby", category: "ruby"},
-  %Reader.Article{url: "http://www.rails.com", title: "A fantastic article about Rails", category: "rails"},
-  %Reader.Article{
-    url: "http://blog.acolyer.org/2015/01/13/unikernels-library-operating-systems-for-the-cloud/",
-    title: "Unikernels: Library Operating Systems for the Cloud | the morning paper",
-    category: "distributed_systems"},
-  %Reader.Article{url: "http://www.abc.com", title: "A fantastic article about hard CS", category: "computer_science"},
-  %Reader.Article{url: "http://www.nbc.com", title: "An article about distributed systems", category: "distributed_systems"},
-  %Reader.Article{url: "http://www.amazon.com", title: "An article about AWS", category: "aws"}
-]
+  categories = ["elixir", "phoenix", "ruby", "rails", "distributed_systems", "otp"]
+  domains    = ["org", "com", "net", "biz", "io"]
 
-for article <- articles, do: Reader.Repo.insert!(article)
-
+  for category <- categories, domain <- domains do
+    %Reader.Article{
+      url: "http://www.#{category}.#{domain}",
+      title: "#{Phoenix.Naming.humanize(category)} #{String.upcase(domain)}",
+      category: category
+    } |> Reader.Repo.insert!
+  end
+end
